Wrap Uint256 instead of alias.
This should help to make it easier to a) not leak ethereum-types
outside, and b) solve the problem where big endian (i.e. MIPS) tests are
not passing.
