+++ Cargo.toml
@@ -7,7 +7,9 @@ license = "Apache 2.0"
 [dependencies]
 num = "0.2.0"
 serde = "1.0.70"
+ethereum-types = "0.3.2"
+
+[dev-dependencies]
 serde_derive = "1.0.70"
 serde_json = "1.0.24"
 lazy_static = "1.0.2"
-ethereum-types = "0.3.2"
